admin/requirements.txt
Django==5.2
django-cors-headers==4.7.0
djangorestframework==3.16.0
pillow==11.1.0
psycopg2-binary==2.9.10
pika==1.3.2